refactor(client): split ModuleGroupSection into focused files to clear import/max-dependencies
